Added original code
Starting point will be the Mandelbrot code I wrote as a freshman at USU. This code generated a bitmap image, was single threaded, and was generally very slow. Goals with this project are to use CMake as a build system, multithread the mandelbrot generation, utilize AVX2, and perhaps output to a window, instead of an image.